server/requirements.txt
Django==1.11.7
requests==2.18.4
psycopg2==2.7.3.2
python-dotenv==0.7.1
djangorestframework==3.7.3
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
django-bootstrap3==9.1.0
